+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## x.y.z (yyyy-mm-dd)
+
+### Bug fixes
+
+- Fail fast if duplicate class names are detected. ([#19](https://github.com/EvidentSolutions/apina/issues/19))
+
 ## 0.3.0 (2015-08-19)
 
 ### New features